+++ src/templates/domain/models/Template.ts
@@ -1,7 +1,7 @@
 import { DatasetMetadataBlock, TermsOfUse } from '../../../datasets/domain/models/Dataset'
 import { License } from '../../../licenses/domain/models/License'
 
-export interface DatasetTemplate {
+export interface Template {
   id: number
   name: string
   collectionAlias: string
@@ -11,13 +11,13 @@ export interface DatasetTemplate {
   createDate: string
   // 👇 From Edit Template Metadata
   datasetMetadataBlocks: DatasetMetadataBlock[]
-  instructions: DatasetTemplateInstruction[]
+  instructions: TemplateInstruction[]
   // 👇 From Edit Template Terms
   termsOfUse: TermsOfUse
   license?: License // This license property is going to be present if not custom terms are added in the UI
 }
 
-export interface DatasetTemplateInstruction {
+export interface TemplateInstruction {
   instructionField: string
   instructionText: string
 }