+++ src/templates/domain/useCases/GetDatasetTemplates.ts
@@ -1,9 +1,9 @@
 import { ROOT_COLLECTION_ID } from '../../../collections/domain/models/Collection'
 import { UseCase } from '../../../core/domain/useCases/UseCase'
-import { CreateDatasetTemplateDTO } from '../dtos/CreateDatasetTemplateDTO'
+import { Template } from '../models/Template'
 import { ITemplatesRepository } from '../repositories/ITemplatesRepository'
 
-export class CreateDatasetTemplate implements UseCase<void> {
+export class GetDatasetTemplates implements UseCase<Template[]> {
   private templatesRepository: ITemplatesRepository
 
   constructor(templatesRepository: ITemplatesRepository) {
@@ -11,17 +11,13 @@ export class CreateDatasetTemplate implements UseCase<void> {
   }
 
   /**
-   * Creates a Dataset Template in the specified collection.
+   * Returns a Template array containing the templates of the requested collection, given the collection identifier or alias.
    *
-   * @param {CreateDatasetTemplateDTO} template - Template definition payload.
    * @param {number | string} [collectionIdOrAlias = ':root'] - A generic collection identifier, which can be either a string (for queries by CollectionAlias), or a number (for queries by CollectionId)
-   * If this parameter is not set, the default value is: ':root'.
-   * @returns {Promise<void>}
+   * If this parameter is not set, the default value is: ':root'
+   * @returns {Promise<Template[]>}
    */
-  async execute(
-    template: CreateDatasetTemplateDTO,
-    collectionIdOrAlias: number | string = ROOT_COLLECTION_ID
-  ): Promise<void> {
-    return await this.templatesRepository.createDatasetTemplate(collectionIdOrAlias, template)
+  async execute(collectionIdOrAlias: number | string = ROOT_COLLECTION_ID): Promise<Template[]> {
+    return await this.templatesRepository.getDatasetTemplates(collectionIdOrAlias)
   }
 }