+++ test/functional/templates/createDatasetTemplate.test.ts
@@ -1,8 +1,8 @@
 import { ApiConfig } from '../../../src'
 import { DataverseApiAuthMechanism } from '../../../src/core/infra/repositories/ApiConfig'
 import { TestConstants } from '../../testHelpers/TestConstants'
-import { createDatasetTemplate, getDatasetTemplates } from '../../../src/template'
-import { CreateDatasetTemplateDTO } from '../../../src/template/domain/dtos/CreateDatasetTemplateDTO'
+import { createTemplate, getDatasetTemplates } from '../../../src/templates'
+import { CreateDatasetTemplateDTO } from '../../../src/templates/domain/dtos/CreateDatasetTemplateDTO'
 import { MetadataFieldTypeClass } from '../../../src/metadataBlocks/domain/models/MetadataBlock'
 import { deleteDatasetTemplateViaApi } from '../../testHelpers/datasets/datasetTemplatesHelper'
 
@@ -47,7 +47,7 @@ describe('CreateTemplate.execute', () => {
         }
       ]
     }
-    await createDatasetTemplate.execute(templateDto)
+    await createTemplate.execute(templateDto)
     const templates = await getDatasetTemplates.execute(':root')
 
     expect(templates[templates.length - 1].name).toBe(templateDto.name)