+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This changelog follows the principles of [Keep a Changelog](https://keepachangel
 
 - New Use Case: [Get Collections For Linking Use Case](./docs/useCases.md#get-collections-for-linking).
 
-- New Use Case: [Create a Dataset Template](./docs/useCases.md#create-a-dataset-template) under Templates.
+- New Use Case: [Create a Template](./docs/useCases.md#create-a-template) under Templates.
 
 - New Use Case: [Get a Template](./docs/useCases.md#get-a-template) under Templates.